Updated some types

+++ device_libraries/firehose_ipu.hpp
@@ -17,9 +17,9 @@
 
 void printMatrix(std::string matrix_name, std::vector<float> matrix, int cols);
 
-void frontEnd_TensorDecomp(bool& flag, int& rows, int& cols, int& exp_size, std::vector<float>& cpu_input0, std::vector<float>& cpu_output0, std::vector<float>& cpu_output1);
+void frontEnd_TensorDecomp(bool& flag, long unsigned int& rows, long unsigned int& cols, long unsigned int& exp_size, std::vector<float>& cpu_input0, std::vector<float>& cpu_output0, std::vector<float>& cpu_output1);
 
-void backEnd_TensorDecomp(poplar::Engine& engine, bool& flag, int& exp_size) {
+void backEnd_TensorDecomp(poplar::Engine& engine, bool& flag, long unsigned int& exp_size) {
 
     for (int i = 0; i < exp_size; i++) {
         while(!flag) {}
